+++ src/main/java/com/greatdistances/self20230324/model/data/EquiplistRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EquiplistRepository extends CrudRepository<Equiplist, Integer> {
+public interface EquiplistRepository extends CrudRepository<Equiplist, Long> {
 }
